+++ tests/e2e/utils/CHANGELOG.md
@@ -15,6 +15,7 @@
 - `addProductToOrder( orderId, productName )` component which adds the provided productName to the passed in orderId
 - `createCoupon( couponAmount )` component which accepts a coupon amount string (it defaults to 5) and creates a basic coupon. Returns the generated coupon code.
 - `evalAndClick( selector )` use Puppeteer page.$eval to select and click and element.
+- `selectOptionInSelect2( selector, value )` util helper method that search and select in any select2 type field
 
 ## Changes